01-generative-ai/gemini-langchain-apps/apps/requirements.txt
langchain-google-genai
streamlit
python-dotenv